requirements.txt
selenium==2.48.0
PyYAML==3.11
